+++ test/projects_test.dart
@@ -91,14 +91,13 @@ void main() {
     test('can install dependencies and compile project', () async {
       final jbResult = await runJb(Directory(withDepsProjectDir), const []);
       expectSuccess(jbResult);
-      // FIXME enable again when we are able to build reproducible jars
-      // await verifyDependenciesChecksums(Directory(withDepsProjectDir), {
-      //   'com.example:lists:1.0': 'd1961a9adfdf3c40afb3064cd23c8647cdeecdc4',
-      //   'org.slf4j:slf4j-api:1.7.36':
-      //       '6c62681a2f655b49963a5983b8b0950a6120ae14',
-      //   'org.slf4j:slf4j-simple:1.7.36':
-      //       'a41f9cfe6faafb2eb83a1c7dd2d0dfd844e2a936',
-      // });
+      await verifyDependenciesChecksums(Directory(withDepsProjectDir), {
+        'com.example:lists:1.0': '47e9f0cce6d813b7e06a0381550e1c2b78292975',
+        'org.slf4j:slf4j-api:1.7.36':
+            '6c62681a2f655b49963a5983b8b0950a6120ae14',
+        'org.slf4j:slf4j-simple:1.7.36':
+            'a41f9cfe6faafb2eb83a1c7dd2d0dfd844e2a936',
+      });
 
       await assertDirectoryContents(
         Directory(p.join(withDepsProjectDir, 'build')),